Ticket 0035: review fix — rename _common.py, remove sys.path hack, add missing-score warning, widen rank-flip range
1. Rename matching/_common.py to matching/common.py (no underscore prefix)
   and update imports in lp.py, phased.py, and README.md.
2. Replace sys.path.insert hack in test_sweep_matching_threshold.py with
   pythonpath = ["."] in pyproject.toml pytest config.
3. Add warnings.warn() in _compute_metrics_at_threshold when a fuzzy row
   lacks similarity_score, making flat F1 lines from old CSVs explicit.
4. Widen _print_rank_stability threshold range from [80, 95] to [75, 95]
   to match the report text ("de 75 a 95").
