+++ pyproject.toml
@@ -42,6 +42,7 @@ ignore = ["E501"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+pythonpath = ["."]
 markers = [
     "integration: spawns subprocesses or uses sleep-based timing",
     "slow: requires network access or real data",
