+++ src/aedist/matching/lp.py
@@ -57,7 +57,7 @@ from pulp import (
 )
 from rapidfuzz import fuzz
 
-from aedist.matching._common import build_result_row
+from aedist.matching.common import build_result_row
 
 # ---------------------------------------------------------------------------
 # Default MILP parameters for record reconciliation
